in-progress: get block operation
