+++ bitswap/interface.go
@@ -1,0 +1,5 @@
+package bitswap
+
+// TODO(brian): use a Bitswap interface. Let the struct be a private
+// implementation. Let the factory method return the struct but
+// expose the interface as its return value
